+++ changes.md
@@ -249,6 +249,7 @@
 <details>
 <summary>Added Enum(s)</summary>
 
+- added enum `mcp-server` to type `ReferenceTypeId`
 - added enum `variant` to type `AttributeReferenceTypeId`
 - added enum `variant` to type `ChangeSubscriptionResourceTypeId`
 - added enum `variant` to type `MessageSubscriptionResourceTypeId`